requirements.txt
discord.py==1.2.0
praw==6.3.1
async_timeout==3.0.1
pyosu==0.5.2
random_cat==1.0.1
requests==2.22.0
geopy==1.20.0
youtube_dl==2020.3.8
SQLAlchemy==1.3.5
Pillow==6.1.0
tzwhere==3.0.3
